requirements.txt
requests
flask
coverage
Authlib
python-dotenv
uuid